# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/clients": {
      "get": {
        "operationId": "AdminClientsController_listClients",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Legacy offset paging. Prefer cursor.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque keyset cursor (R-2.4).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "withTotal",
            "required": false,
            "in": "query",
            "description": "Counting is a full scan.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search email and name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "individual",
                "referral",
                "partner"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "active",
                "pending",
                "suspended"
              ],
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                0,
                1
              ],
              "type": "number"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "description": "Exact match on the country tag.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "description": "Tag SLUG, not id (ADM-14).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "createdAt",
                "email",
                "firstName",
                "status",
                "type",
                "verificationLevel",
                "country"
              ],
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Paginated, filterable, sortable client list",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/clients/export": {
      "get": {
        "description": "Takes the SAME filters as GET /admin/clients and covers every matching row rather than one page. Paging parameters are not accepted. Client scope and field masking apply exactly as they do to the list.",
        "operationId": "AdminClientsController_exportClients",
        "parameters": [
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "csv"
              ],
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Search email and name.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "individual",
                "referral",
                "partner"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "active",
                "pending",
                "suspended"
              ],
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                0,
                1
              ],
              "type": "number"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "description": "Exact match on the country tag.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "description": "Tag SLUG, not id (ADM-14).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "createdAt",
                "email",
                "firstName",
                "status",
                "type",
                "verificationLevel",
                "country"
              ],
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A CSV file. `Content-Disposition` names it `clients-<YYYY-MM-DD>.csv`.",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Export the filtered client list as CSV",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/clients/{id}": {
      "get": {
        "description": "Each section is gated on its own permission and is ABSENT when the caller lacks it — which is deliberately different from present-and-empty, so a screen can distinguish \"hidden from you\" from \"this client has none\".",
        "operationId": "AdminClientsController_getClientProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "A client's full profile — KYC, documents, trading accounts, referrals (FR-ADM-01)",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/clients/{id}/status": {
      "patch": {
        "operationId": "AdminClientsController_setClientStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Suspend or reactivate a client account (requires users.suspend)",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/kyc/{userId}": {
      "get": {
        "operationId": "AdminComplianceController_getKyc",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycSubmissionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Get full KYC submission for a user",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "OxShare CRM API",
    "description": "Forex/CFD Introducing-Broker CRM — Phase 1",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oxshare_crm_portal_at",
        "description": "Portal session. Set by POST /v1/auth/login; httpOnly, so not settable here."
      },
      "admin": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oxshare_crm_admin_at",
        "description": "Admin session. Set by POST /v1/admin/auth/login; httpOnly, so not settable here."
      }
    },
    "schemas": {
...
      "ClientTagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "Stable machine name. Filter with ?tag=<slug>; a rename does not change it."
          },
          "label": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "label",
          "createdAt"
        ]
      },
      "ClientRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "individual",
              "referral",
              "partner"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "pending",
              "suspended"
            ]
          },
          "verificationLevel": {
            "type": "number",
            "enum": [
              0,
              1
            ]
          },
          "country": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientTagDto"
            }
          }
        },
        "required": [
          "id",
          "type",
          "status",
          "verificationLevel"
        ]
      },
      "ClientListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientRowDto"
            }
          },
          "maskedFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "total": {
            "type": "number",
            "description": "Only when ?withTotal=true. Counting 219,000 rows is a full scan."
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "maskedFields",
          "nextCursor",
          "page",
          "limit"
        ]
      },
      "ProfileKycDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "none",
              "pending",
              "in_review",
              "approved",
              "rejected"
            ]
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string"
          },
          "reviewedAt": {
            "format": "date-time",
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          },
          "documentCount": {
            "type": "number",
            "description": "How many documents the submission carries."
          }
        },
        "required": [
          "status",
          "documentCount"
        ]
      },
      "ProfileTradingAccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "mt5Login": {
            "type": "string"
          },
          "mt5Group": {
            "type": "string"
          },
          "environment": {
            "type": "string",
            "enum": [
              "live",
              "demo"
            ]
          },
          "tier": {
            "type": "string"
          },
          "leverage": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "mt5Login",
          "environment",
          "createdAt"
        ]
      },
      "ProfileReferrerDto": {
        "type": "object",
        "properties": {
          "ibUserId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "description": "False when the attribution was switched off."
          },
          "since": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "ibUserId",
          "email",
          "firstName",
          "lastName",
          "active",
          "since"
        ]
      },
      "ProfileReferredClientDto": {
        "type": "object",
        "properties": {
          "clientUserId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "since": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "clientUserId",
          "email",
          "firstName",
          "lastName",
          "active",
          "since"
        ]
      },
      "ClientProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "individual",
              "referral",
              "partner"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "pending",
              "suspended"
            ]
          },
          "verificationLevel": {
            "type": "number",
            "enum": [
              0,
              1
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "country": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClientTagDto"
            }
          },
          "kyc": {
            "description": "Absent without kyc.view.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProfileKycDto"
              }
            ]
          },
          "documents": {
            "description": "Document filenames. Absent without kyc.documents.view.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tradingAccounts": {
            "description": "Absent without trading.view.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileTradingAccountDto"
            }
          },
          "referrer": {
            "description": "Absent without partners.view.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProfileReferrerDto"
              }
            ]
          },
          "referredClients": {
            "description": "Capped — see referredTotal. Absent without partners.view.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileReferredClientDto"
            }
          },
          "referredShown": {
            "type": "number",
            "description": "How many referredClients were returned; the list is capped for one screen."
          },
          "maskedFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "type",
          "status",
          "verificationLevel",
          "emailVerified",
          "tags",
          "maskedFields"
        ]
      },
      "ClientStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended"
            ],
            "description": "Suspending blocks sign-in but preserves the client and their ledger history."
          }
        },
        "required": [
          "status"
        ]
      },
...
      "KycDocumentDto": {
        "type": "object",
        "properties": {
          "docType": {
            "type": "string"
          },
          "frontFilePath": {
            "type": "string"
          },
          "backFilePath": {
            "type": "string"
          },
          "frontFileName": {
            "type": "string"
          },
          "backFileName": {
            "type": "string"
          }
        }
      },
      "KycSelfieDto": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          }
        }
      },
      "KycAddressProofDto": {
        "type": "object",
        "properties": {
          "docType": {
            "type": "string"
          },
          "filePath": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "page2FilePath": {
            "type": "string"
          },
          "page2FileName": {
            "type": "string"
          }
        }
      },
      "KycUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "individual",
              "referral",
              "partner"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "pending",
              "suspended"
            ]
          },
          "verificationLevel": {
            "type": "number",
            "enum": [
              0,
              1
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "country": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName"
        ]
      },
      "KycSubmissionDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "not_started",
              "in_progress",
              "submitted",
              "under_review",
              "approved",
              "rejected"
            ]
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string"
          },
          "reviewedAt": {
            "format": "date-time",
            "type": "string"
          },
          "reviewedBy": {
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          },
          "rejectedFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "personalInfo": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "document": {
            "$ref": "#/components/schemas/KycDocumentDto"
          },
          "selfie": {
            "$ref": "#/components/schemas/KycSelfieDto"
          },
          "addressProof": {
            "$ref": "#/components/schemas/KycAddressProofDto"
          },
          "user": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/KycUserDto"
              }
            ]
          }
        },
        "required": [
          "userId",
          "status"
        ]
      },
...
    }
  }
}